Fix crashes
